requirements.txt
# Core dependencies
pan-os-python>=1.3.0

# CLI and utilities
argparse
python-dateutil>=2.8.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.0
pytest-cov>=4.0
mock>=5.0

# Utilities
rich>=13.0
python-json-logger>=2.0.0